requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-login==0.5.0
werkzeug==2.0.1
paddleocr==2.6.0
paddlepaddle==2.4.2
torch==2.4.1
torchvision==0.19.1
numpy==1.21.2
pillow==8.3.2
opencv-python==4.5.3.56
spacy==3.1.3
nltk==3.6.3
gtts==2.2.3
apscheduler==3.8.1
scikit-learn==0.24.2
matplotlib==3.4.3
tqdm==4.62.2 